requirements.txt
fastapi
uvicorn[standard]
groq
httpx
playwright==1.55.0
pdfplumber
Pillow
google.generativeai
python-dotenv
pandas